Rename the `crl` module to `revocation_checking`.
In preparation for support of additional revocation checking strategies,
first one being support for OCSP stapling.
